requirements.txt
starlette>=0.36.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pyyaml>=6.0
python-telegram-bot>=21.0
